:recycle: Decouple SchemaDefinitionService from FhirDefinitionService
